Add rawNeurons attribute to Layers class
Store raw values, i.e. neurons without activation applied, in every
layer. This allows custom activation functions to be passed.

The backpropagation algorithms requires these values for the "error
term" to be computed. The error term relies on the derivative of
a neurons output (applied activation) with respect to the raw
output (activation not applied). This raw output will be saved in
the array rawNeurons.